requirements.txt
PyQt5>=5.15.0
mss>=10.0.0
Pillow>=10.0.0
requests>=2.31.0
PyJWT>=2.8.0
pytz>=2024.1
websocket-client>=1.8.0
pywin32>=306; sys_platform == 'win32'
psutil>=5.9.0
